Add Custom Exceptions

+++ src/Core/MCL.Application/Features/LeaveAllocations/Handlers/Commands/DeleteLeaveAllocationCommandHandler.cs
@@ -1,4 +1,6 @@
-﻿namespace MCL.Application.Features.LeaveAllocations.Handlers.Commands;
+﻿using MCL.Domain;
+
+namespace MCL.Application.Features.LeaveAllocations.Handlers.Commands;
 public class DeleteLeaveAllocationCommandHandler : IRequestHandler<DeleteLeaveAllocationCommand, Unit>
 {
     public DeleteLeaveAllocationCommandHandler(ILeaveAllocationRepository repository)
@@ -10,6 +12,7 @@ public class DeleteLeaveAllocationCommandHandler : IRequestHandler<DeleteLeaveAl
     public async Task<Unit> Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
     {
         var leaveAllocation = await Repository.GetAsync(request.Id, cancellationToken);
+        if (leaveAllocation is null) throw new Exceptions.NotFoundException(nameof(LeaveAllocation), request.Id);
         await Repository.DeleteAsync(leaveAllocation, cancellationToken);
         return Unit.Value;
     }
