+++ src/Core/MCL.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,4 +1,6 @@
-﻿namespace MCL.Application.Features.LeaveRequests.Handlers.Commands;
+﻿using MCL.Domain;
+
+namespace MCL.Application.Features.LeaveRequests.Handlers.Commands;
 public class DeleteLeaveRequestCommandHandler : IRequestHandler<DeleteLeaveRequestCommand, Unit>
 {
     public DeleteLeaveRequestCommandHandler(ILeaveRequestRepository repository)
@@ -10,6 +12,8 @@ public class DeleteLeaveRequestCommandHandler : IRequestHandler<DeleteLeaveReque
     public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
     {
         var leaveRequest = await Repository.GetAsync(request.Id, cancellationToken);
+        if (leaveRequest is null) throw new Exceptions.NotFoundException(nameof(LeaveRequest), request.Id);
+        
         await Repository.DeleteAsync(leaveRequest, cancellationToken);
         return Unit.Value;
     }
